0905/requirements.txt
requests

python-dotenv

google-generativeai
